+++ .phan/plugins/UnknownElementTypePlugin.php
@@ -1,9 +1,14 @@
 <?php declare(strict_types=1);
 
 use Phan\CodeBase;
+use Phan\Language\Element\AddressableElement;
 use Phan\Language\Element\Func;
 use Phan\Language\Element\Method;
 use Phan\Language\Element\Property;
+use Phan\Language\Type;
+use Phan\Language\Type\ArrayType;
+use Phan\Language\Type\NullType;
+use Phan\Language\UnionType;
 use Phan\PluginV2;
 use Phan\PluginV2\AnalyzeFunctionCapability;
 use Phan\PluginV2\AnalyzeMethodCapability;
@@ -17,6 +22,19 @@ class UnknownElementTypePlugin extends PluginV2 implements
     AnalyzeMethodCapability,
     AnalyzePropertyCapability
 {
+
+    /**
+     * Returns true for array, ?array, and array|null
+     */
+    private static function isRegularArray(UnionType $type) : bool
+    {
+        return $type->hasTypeMatchingCallback(static function (Type $type) : bool {
+            return get_class($type) === ArrayType::class;
+        }) && !$type->hasTypeMatchingCallback(static function (Type $type) : bool {
+            return get_class($type) !== ArrayType::class && !($type instanceof NullType);
+        });
+    }
+
     /**
      * @param CodeBase $code_base
      * The code base in which the method exists
@@ -35,16 +53,16 @@ class UnknownElementTypePlugin extends PluginV2 implements
         if ($method->getFQSEN() !== $method->getRealDefiningFQSEN()) {
             return;
         }
+
+        $this->performChecks(
+            $code_base,
+            $method,
+            'PhanPluginUnknownMethodReturnType',
+            'Method {METHOD} has no declared or inferred return type',
+            'PhanPluginUnknownArrayMethodReturnType',
+            'Method {METHOD} has a return type of array, but does not specify any key types or value types'
+        );
         // NOTE: Placeholders can be found in \Phan\Issue::uncolored_format_string_for_replace
-        if ($method->getUnionType()->isEmpty()) {
-            $this->emitIssue(
-                $code_base,
-                $method->getContext(),
-                'PhanPluginUnknownMethodReturnType',
-                'Method {METHOD} has no declared or inferred return type',
-                [(string)$method->getFQSEN()]
-            );
-        }
         foreach ($method->getParameterList() as $parameter) {
             if ($parameter->getUnionType()->isEmpty()) {
                 $this->emitIssue(
@@ -55,6 +73,34 @@ class UnknownElementTypePlugin extends PluginV2 implements
                     [(string)$method->getFQSEN(), $parameter->getName()]
                 );
             }
+        }
+    }
+
+    private function performChecks(
+        CodeBase $code_base,
+        AddressableElement $element,
+        string $issue_type_for_empty,
+        string $message_for_empty,
+        string $issue_type_for_unknown_array,
+        string $message_for_unknown_array
+    ) {
+        $union_type = $element->getUnionType();
+        if ($union_type->isEmpty()) {
+            $this->emitIssue(
+                $code_base,
+                $element->getContext(),
+                $issue_type_for_empty,
+                $message_for_empty,
+                [(string)$element->getFQSEN()]
+            );
+        } elseif (self::isRegularArray($union_type)) {
+            $this->emitIssue(
+                $code_base,
+                $element->getContext(),
+                $issue_type_for_unknown_array,
+                $message_for_unknown_array,
+                [(string)$element->getFQSEN()]
+            );
         }
     }
 
@@ -81,6 +127,21 @@ class UnknownElementTypePlugin extends PluginV2 implements
             } else {
                 $issue = 'PhanPluginUnknownFunctionReturnType';
                 $message = 'Function {FUNCTION} has no declared or inferred return type';
+            }
+            $this->emitIssue(
+                $code_base,
+                $function->getContext(),
+                $issue,
+                $message,
+                [(string)$function->getFQSEN()]
+            );
+        } elseif (self::isRegularArray($function->getUnionType())) {
+            if ($function->getFQSEN()->isClosure()) {
+                $issue = 'PhanPluginUnknownArrayClosureReturnType';
+                $message = 'Closure {FUNCTION} has a return type of array, but does not specify key or value types';
+            } else {
+                $issue = 'PhanPluginUnknownArrayFunctionReturnType';
+                $message = 'Function {FUNCTION} has a return type of array, but does not specify key or value types';
             }
             $this->emitIssue(
                 $code_base,
@@ -128,15 +189,14 @@ class UnknownElementTypePlugin extends PluginV2 implements
         if ($property->getFQSEN() !== $property->getRealDefiningFQSEN()) {
             return;
         }
-        if ($property->getUnionType()->isEmpty()) {
-            $this->emitIssue(
-                $code_base,
-                $property->getContext(),
-                'PhanPluginUnknownPropertyType',
-                "Property {PROPERTY} has an initial type that cannot be inferred",
-                [(string)$property->getFQSEN()]
-            );
-        }
+        $this->performChecks(
+            $code_base,
+            $property,
+            'PhanPluginUnknownPropertyType',
+            'Property {PROPERTY} has an initial type that cannot be inferred',
+            'PhanPluginUnknownArrayPropertyType',
+            'Property {PROPERTY} has an array type, but does not specify any key types or value types'
+        );
     }
 }