+++ src/phan.php
@@ -25,6 +25,7 @@ $cli = CLI::fromArgv();
 $is_issue_found =
     Phan::analyzeFileList(
         $code_base,
+        /** @return array<int,string> */
         static function (bool $recompute_file_list = false) use ($cli) : array {
             if ($recompute_file_list) {
                 $cli->recomputeFileList();